fixed naming scheme so that the main view is now titled main activity, removed default title bar at the top of the screen, adjusted margins, added scroll bar, added static title bar on chat room list
