--- app/src/main/res/layout/activity_main_view.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainView">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/chat_list_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#0A0A0A"
-        android:padding="5dp"
-        android:scrollbars="vertical" />
-
-
-</RelativeLayout>
