+++ app/src/main/res/layout/chat_card.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_marginBottom="-1dp"
@@ -10,7 +11,8 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_margin="1dp"
+
+        android:layout_marginTop="1dp"
         android:background="#0A0A0A">
 
         <TextView
@@ -20,13 +22,12 @@
             android:layout_alignParentStart="true"
             android:layout_alignParentLeft="true"
             android:layout_alignParentTop="true"
-            android:layout_marginStart="50dp"
-            android:layout_marginLeft="50dp"
+            android:layout_marginStart="70dp"
+            android:layout_marginLeft="70dp"
             android:layout_marginTop="10dp"
-            android:text="Temp text"
+            android:text="@string/chat_name"
             android:textColor="#FFFFFF"
-            android:textSize="20sp"
-            android:textStyle="bold" />
+            android:textSize="20sp" />
 
         <TextView
             android:id="@+id/mes_prev_text"
@@ -36,10 +37,10 @@
             android:layout_alignParentLeft="true"
 
             android:layout_alignParentTop="true"
-            android:layout_marginStart="50dp"
-            android:layout_marginLeft="50dp"
+            android:layout_marginStart="70dp"
+            android:layout_marginLeft="70dp"
             android:layout_marginTop="30dp"
-            android:text="Temp text 2"
+            android:text="@string/previous_message"
             android:textColor="#AAAAAA"
             android:textSize="15sp" />
 
@@ -51,12 +52,27 @@
             android:layout_alignParentLeft="true"
 
             android:layout_alignParentTop="true"
-            android:layout_marginStart="50dp"
-            android:layout_marginLeft="50dp"
+            android:layout_marginStart="70dp"
+            android:layout_marginLeft="70dp"
             android:layout_marginTop="47dp"
             android:layout_marginBottom="10dp"
-            android:text="Temp date"
+            android:text="@string/last_date"
             android:textColor="#646464"
-            android:textSize="10sp" />
+            android:textSize="12sp" />
+
+        <ImageView
+            android:id="@+id/chatIcon"
+            android:layout_width="64dp"
+            android:layout_height="67dp"
+            android:layout_alignParentStart="true"
+            android:layout_alignParentLeft="true"
+            android:layout_marginStart="3dp"
+            android:layout_marginLeft="3dp"
+            android:layout_marginTop="3dp"
+            android:layout_marginRight="3dp"
+            android:layout_marginBottom="0dp"
+            android:contentDescription="@string/stock_image_for_the_chat_room_icon"
+            app:srcCompat="@drawable/ic_launcher_foreground"
+            tools:ignore="VectorDrawableCompat" />
     </RelativeLayout>
 </androidx.cardview.widget.CardView>
