+++ app/src/main/res/layout/chat_room.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
@@ -76,6 +75,7 @@
         android:layout_marginLeft="0dp"
         android:layout_marginTop="19dp"
         android:background="@drawable/abc_vector_test"
-        android:textAllCaps="false" />
+        android:textAllCaps="false"
+        tools:ignore="PrivateResource" />
 
 </RelativeLayout>
